+++ Tests/Unit/Validation/Fixtures/TestingValidatableModel.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OliverKlee\Oelib\Tests\Unit\Validation\Fixtures;
+
+use TYPO3\CMS\Extbase\DomainObject\AbstractEntity;
+
+final class TestingValidatableModel extends AbstractEntity
+{
+    /**
+     * @var string
+     */
+    protected $title;
+
+    public function __construct(string $title = '')
+    {
+        $this->title = $title;
+    }
+
+    public function getTitle(): string
+    {
+        return $this->title;
+    }
+}